refactor(control-plane): move DEFAULT_*_API_KEY fallback from repo to service layer
The in-memory storage (memory.rs) was missing the DEFAULT_*_API_KEY
environment variable fallback that exists in the PostgreSQL
implementation (repositories.rs).

Rather than duplicate the logic, move it to the service layer
(LlmResolverService) where it belongs architecturally - repositories
should only handle data storage/retrieval, not business logic like
environment variable fallbacks.

Changes:
- Add get_default_api_key_from_env() to llm_resolver.rs
- Apply env fallback after provider lookup in resolve_model/resolve_default_model
- Remove fallback logic from repositories.rs get_provider_with_api_key()
- Move tests from repositories.rs to llm_resolver.rs
