+++ crates/control-plane/src/services/llm_resolver.rs
@@ -2,11 +2,42 @@
 //
 // This service handles the resolution of LLM models with their provider credentials,
 // including API key decryption. Used by gRPC service for worker communication.
+//
+// API key resolution order (handled at service layer):
+// 1. Decrypted key from database (if set)
+// 2. Environment variable fallback (for development convenience):
+//    - openai: DEFAULT_OPENAI_API_KEY
+//    - anthropic: DEFAULT_ANTHROPIC_API_KEY
 
 use crate::storage::{EncryptionService, StorageBackend};
 use anyhow::{anyhow, Result};
 use std::sync::Arc;
 use uuid::Uuid;
+
+/// Get default API key from environment variable based on provider type.
+///
+/// Environment variables (for development convenience):
+/// - DEFAULT_OPENAI_API_KEY: Fallback API key for OpenAI providers
+/// - DEFAULT_ANTHROPIC_API_KEY: Fallback API key for Anthropic providers
+///
+/// These are only used when the provider doesn't have an API key set in the database.
+pub fn get_default_api_key_from_env(provider_type: &str) -> Option<String> {
+    get_default_api_key_with_lookup(provider_type, |name| std::env::var(name).ok())
+}
+
+/// Testable version with injectable env lookup.
+fn get_default_api_key_with_lookup<F>(provider_type: &str, env_lookup: F) -> Option<String>
+where
+    F: Fn(&str) -> Option<String>,
+{
+    let env_var = match provider_type.to_lowercase().as_str() {
+        "openai" => "DEFAULT_OPENAI_API_KEY",
+        "anthropic" => "DEFAULT_ANTHROPIC_API_KEY",
+        _ => return None,
+    };
+
+    env_lookup(env_var).filter(|s| !s.is_empty())
+}
 
 /// Resolved model with provider credentials (decrypted API key)
 ///
@@ -57,15 +88,20 @@ impl LlmResolverService {
             None => return Ok(None),
         };
 
-        // Decrypt the API key
+        // Decrypt the API key from database
         let provider_with_key = self
             .db
             .get_provider_with_api_key(&provider_row, &encryption)?;
 
+        // Apply env fallback if no API key in database
+        let api_key = provider_with_key
+            .api_key
+            .or_else(|| get_default_api_key_from_env(&provider_with_key.provider_type));
+
         Ok(Some(ResolvedModel {
             model_id: model_row.model_id,
             provider_type: provider_with_key.provider_type,
-            api_key: provider_with_key.api_key,
+            api_key,
             base_url: provider_with_key.base_url,
         }))
     }
@@ -93,15 +129,20 @@ impl LlmResolverService {
             None => return Ok(None),
         };
 
-        // Decrypt the API key
+        // Decrypt the API key from database
         let provider_with_key = self
             .db
             .get_provider_with_api_key(&provider_row, &encryption)?;
 
+        // Apply env fallback if no API key in database
+        let api_key = provider_with_key
+            .api_key
+            .or_else(|| get_default_api_key_from_env(&provider_with_key.provider_type));
+
         Ok(Some(ResolvedModel {
             model_id: model_row.model_id,
             provider_type: provider_with_key.provider_type,
-            api_key: provider_with_key.api_key,
+            api_key,
             base_url: provider_with_key.base_url,
         }))
     }
@@ -109,5 +150,78 @@ impl LlmResolverService {
     /// Check if encryption service is available
     pub fn has_encryption(&self) -> bool {
         self.encryption.is_some()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn mock_env<'a>(vars: &'a [(&'a str, &'a str)]) -> impl Fn(&str) -> Option<String> + 'a {
+        move |name| {
+            vars.iter()
+                .find(|(k, _)| *k == name)
+                .map(|(_, v)| v.to_string())
+        }
+    }
+
+    #[test]
+    fn test_get_default_api_key_openai() {
+        // Not set
+        assert_eq!(
+            get_default_api_key_with_lookup("openai", mock_env(&[])),
+            None
+        );
+        assert_eq!(
+            get_default_api_key_with_lookup("OpenAI", mock_env(&[])),
+            None
+        );
+
+        // Set
+        let env = mock_env(&[("DEFAULT_OPENAI_API_KEY", "sk-test-key")]);
+        assert_eq!(
+            get_default_api_key_with_lookup("openai", &env),
+            Some("sk-test-key".to_string())
+        );
+        assert_eq!(
+            get_default_api_key_with_lookup("OpenAI", &env),
+            Some("sk-test-key".to_string())
+        );
+    }
+
+    #[test]
+    fn test_get_default_api_key_anthropic() {
+        // Not set
+        assert_eq!(
+            get_default_api_key_with_lookup("anthropic", mock_env(&[])),
+            None
+        );
+
+        // Set
+        let env = mock_env(&[("DEFAULT_ANTHROPIC_API_KEY", "sk-ant-test-key")]);
+        assert_eq!(
+            get_default_api_key_with_lookup("anthropic", &env),
+            Some("sk-ant-test-key".to_string())
+        );
+        assert_eq!(
+            get_default_api_key_with_lookup("Anthropic", &env),
+            Some("sk-ant-test-key".to_string())
+        );
+    }
+
+    #[test]
+    fn test_get_default_api_key_unknown_provider() {
+        let env = mock_env(&[
+            ("DEFAULT_OPENAI_API_KEY", "sk-test"),
+            ("DEFAULT_ANTHROPIC_API_KEY", "sk-ant-test"),
+        ]);
+        assert_eq!(get_default_api_key_with_lookup("azure_openai", &env), None);
+        assert_eq!(get_default_api_key_with_lookup("unknown", &env), None);
+    }
+
+    #[test]
+    fn test_get_default_api_key_empty_value() {
+        let env = mock_env(&[("DEFAULT_OPENAI_API_KEY", "")]);
+        assert_eq!(get_default_api_key_with_lookup("openai", &env), None);
     }
 }
